fix: Rename Network.framework to IMNetwork.framework
Due to this issue https://forums.developer.apple.com/thread/84068 we was able to rename out local network framework.
